Split mux in h1 and h2 files
